requirements.txt
click==6.7
pylint==2.1.1
request==2.19.1

